corrected some bugs and added acknowledgements to README

+++ ft_printf/include/printf.h
@@ -6,7 +6,7 @@
 /*   By: mdesalle <mdesalle@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/02 09:02:17 by mdesalle          #+#    #+#             */
-/*   Updated: 2021/01/10 18:10:16 by mdesalle         ###   ########.fr       */
+/*   Updated: 2021/01/11 10:56:42 by mdesalle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,7 +108,9 @@ char			*ft_strchr(const char *s, int c);
 ** ft_utilities_three.c
 */
 int				ft_baselen(unsigned long x);
-void			ft_putnbr_unsigned(unsigned int nbr, unsigned int m, t_list *box);
-void			ft_putstr_hexa(char *str, long x, t_list *box);
+void			ft_putnbr_unsigned(unsigned int nbr,
+				unsigned int m, t_list *box);
+void			ft_putstr_hexa(char *str,
+				unsigned long long x, t_list *box);
 
 #endif
